ios-angry-fog/requirements.txt
flask
flask-cors
flask-sqlalchemy
requests
python-dotenv
python-multipart
Pillow
requests
